Add external link icon to game rules button and implement GameBoard component (#15)
* ゲームルールボタンに外部リンクアイコンを追加し、ボタンのスタイルを調整

* GameBoardコンポーネントを新規作成し、ゲームの状態管理とUIを実装

* ゲームサーバー接続の状態に応じてローディングとエラーメッセージを表示する機能を追加
